+++ glue/clients/tests/test_scatter_client.py
@@ -17,7 +17,7 @@ from glue.core.component import Component, CategoricalComponent
 from glue.core.data_collection import DataCollection
 from glue.core.data import Data
 from glue.core.roi import RectangularROI, XRangeROI, YRangeROI
-from glue.core.subset import RangeSubsetState, CategoricalRoiSubsetState, AndState
+from glue.core.subset import RangeSubsetState, CategoricalROISubsetState, AndState, OrState
 from glue.tests import example_data
 
 from ..scatter_client import ScatterClient
@@ -698,7 +698,7 @@ class TestCategoricalScatterClient(TestScatterClient):
         roi.set_range(1, 2)
         self.client.apply_roi(roi)
         assert isinstance(data.edit_subset.subset_state,
-                          CategoricalRoiSubsetState)
+                          CategoricalROISubsetState)
         assert data.edit_subset.subset_state.att == self.client.xatt
 
         roi = YRangeROI()
@@ -711,7 +711,7 @@ class TestCategoricalScatterClient(TestScatterClient):
 
         self.client.apply_roi(roi)
         assert isinstance(data.edit_subset.subset_state,
-                          AndState)
+                          OrState)
 
     @pytest.mark.parametrize(('roi_limits', 'mask'), [((0, -0.1, 10, 0.1), [0, 0, 0]),
                                                       ((0, 0.9, 10, 1.1), [1, 0, 0]),